Chapter 1.7 Procedures
Reminder: the stack grows downwards in memory, meaning decreasing the stack
pointer uses up more stack space not less.

The call instruction performs the following: (1) computes the return point (i.e.
the address of the instruction following the call instruc- tion), (2) decrements
the value of %esp by 4, (3) saves the return point at 0(%esp), then (4) directs
the execution to the target of the call.

The ret instruction performs the following: (1) it loads the return point
address from 0(%esp), (2) increments the value of %esp by 4, then (3) directs
the execution to the return point.

In a call, rsp starts out pointing to the return address of the call.

Regarding earlier confusion,
    not sure about this line:
        (emit "  mov [rsp + 8], rsp") ; stack base argument
    tutorial has: movl 4(%esp), %esp

It looks like the tutorial code would pick the stack_base argument directly
off the stack - as it was the last local in main C function, and could
be expected to be in the machine word above the current stack pointer.

In 64-bit calling convention, this argument is passed in rdi.

stack_base is passed via the calling convention, and as it's the  only
argument, it is passed in the rdi register.

So this line has been changed to the following:
    (emit "  mov rsp, rdi")

Using 64-bit cells, Bus Error would occur in deeply nested procedures tests
due to use of rsp pointing to out of bounds memory, need more memory...

   int stack_size = (16 * 4096); /* holds 16K cells */

Increasing 16 to 32768 enables those tests to pass, allowing enough stack size:
   5000000 (call depth in test) * 2 (lambdas) * 8 (wordsize) = 80,000,000
   32768 * 4096 = 134,217,728

Tip: to get gdb to show intel syntax,
% cat ~/.gdbinit
set disassembly-flavor intel

Tip: to see code in gdb,
layout asm
layout reg

The file tests-1.7-req.scm appeared to have a repeat of earlier binary
primitives tests, so these have been removed.

Add gcc flags, -g to include debugging symbols, -fomit-frame-pointer so gcc
does not generate code that uses the rbp register,
-Wall and -pedantic to provide C language warnings

In this implementation, letrec cannot be nested and has to be the outermost
expression.

Remove reliance on ctype.h for isspace, iscntrl (down the track startup.c can be
ported to assembly)

Emit comments next to some assembly code to help debug Bus Error (see above)

passed all 400 tests