Chapter 1.8 Iteration via Proper Tail Calls
Collapsing the stack does not consider how many locals are on it,
simply move the current set of arguments adjacent to the first cell containing
the return address.

tests-1.8 appear to have tests not related to tail calls so these tests have been
removed.

A lot of code can be refactored now

It seems like everything aside from argument evaluation is in a tail
position in this compiler?

passed all 419 tests